Move to device name based layouts (WARNING)
This commit makes devices specifiable in the action map with a path that is based on the device name instead of a specific device+GUID.

This allows input bindings to be generic and apply to all "Fanatec Wheel" devices instead of the specific GUID of your physical device. However, this commit causes issues upon domain reload as the device cannot be replaced by Unity and is forcefully destroyed and re-created, bindings should still function after the domain reload but other things like the input inspector will need to be re-opened as the device has been destroyed and a new one created.

Therefore, this commit has been identified as unsuitable for the main branch and an alternative is to be developed.